+++ apps/Event/groupAdmin/Decrease.js
@@ -31,13 +31,13 @@ class UCDecreaseEvent extends UCGAPlugin {
   async dealNotice() {
     if (!this.Cfg.isNotice) return
     const notice = this.B ? '黑名单' : '群员'
-    const replyMsg = [`[${notice}退群通知]`]
+    const replyMsg = [`【${notice}退群通知】`]
+    replyMsg.push(segment.image(common.getAvatarUrl(this.groupId, 'group')))
     replyMsg.push('\n群聊：' + this.groupName)
     replyMsg.push('\n群号：' + this.groupId)
-    replyMsg.push(segment.image(common.getAvatarUrl(this.groupId, 'group')))
+    replyMsg.push(segment.image(common.getAvatarUrl(this.userId)))
     replyMsg.push(notice + '：' + this.userName)
     replyMsg.push('\n' + notice + 'QQ：' + this.userId)
-    replyMsg.push(segment.image(common.getAvatarUrl(this.userId)))
     await common.sendMsgTo(this.GAconfig.permission?.Master[0] ?? UCPr.GlobalMaster[0], replyMsg, 'Private')
   }