+++ apps/chuoyichuo.js
@@ -3,6 +3,11 @@ import { UCPlugin } from '../models/index.js'
 import { segment } from 'icqq'
 import _ from 'lodash'
 
+/** 戳一戳文本内容，路径UC-plugin/resources/data/chuoyichuo.txt，可自行增删 */
+const textList = file.readFileSync(Path.get('resdata', 'chuoyichuo.txt')).split('\n').map(v => v.trim())
+
+const CD = {}
+
 export default class UCChuoyichuo extends UCPlugin {
   constructor(e) {
     super({
@@ -12,7 +17,18 @@ export default class UCChuoyichuo extends UCPlugin {
       Cfg: 'config.chuoyichuo',
       event: 'notice.group.poke'
     })
-    this.redisData = '[UC]chuoyichuo'
+    if (!e) return
+    this.redisData = '[UC]chuoyichuo:' + this.groupId
+  }
+
+  isCD() {
+    if (!this.Cfg.CD) return false
+    if (CD[this.groupId]) return true
+    CD[this.groupId] = true
+    setTimeout(() => {
+      delete CD[this.groupId]
+    }, this.Cfg.CD * 1000)
+    return false
   }
 
   async accept(e) {
@@ -20,9 +36,10 @@ export default class UCChuoyichuo extends UCPlugin {
     if (!Cfg.isOpen) return false
     if (!this.verifyLevel()) return false
     if (e.target_id !== this.qq) return false
+    if (this.isCD) return false
     log.whiteblod('戳一戳生效')
-    let count = await Data.redisGet(this.redisData + this.groupId, 0)
-    Data.redisSet(this.redisData + this.groupId, ++count, UCDate.EXsecondes)
+    let count = await Data.redisGet(this.redisData, 0)
+    Data.redisSet(this.redisData, ++count, UCDate.EXsecondes)
     if (Cfg.isAutoSetCard) {
       if (!(UCPr.wz.ing && UCPr.wz.groupId === this.groupId)) {
         e.group.setCard(this.qq, `${Cfg.picPath}|${Cfg.groupCard.replace('num', count)}`)
@@ -61,7 +78,7 @@ export default class UCChuoyichuo extends UCPlugin {
         await common.sleep(1.5)
         await this.reply('戳！！')
         await common.sleep(1)
-        return e.group.muteMember(this.userId, Cfg.MuteTime * 60)
+        return e.group.muteMember(this.userId, Cfg.muteTime * 60)
       }
     }
     // 反击
@@ -91,87 +108,10 @@ class UCChuoyichuoSwitchPicPath extends UCPlugin {
 
   async switchPicPath() {
     if (!this.M) return false
-    const picPaths = file.readdirSync(Path.chuoyichuo, { removes: '一键重命名.bat' })
+    const picPaths = file.readdirSync(Path.chuoyichuo, { type: 'Directory' })
     return this.reply(`可选的戳一戳图包：\n${Data.makeArrStr(picPaths)}\n当前使用：${this.Cfg.picPath}\n使用#UC设置戳一戳图包+图包名 切换图包`)
   }
 
 }
 
 UCPr.EventInit(UCChuoyichuoSwitchPicPath)
-
-// 回复文字列表
-const textList = [
-  '看我超级无敌旋风！',
-  '诅咒你买方便面没有叉子！',
-  '救命啊，有变态\n>_<！！！',
-  '哼~~~',
-  '你戳谁呢！你戳谁呢！！！\no(´^｀)o',
-  '是不是要我揍你一顿才开心啊！！！',
-  '食不食油饼？',
-  '不要再戳了！我真的要被你气死了！！！',
-  '怎么会有你这么无聊的人啊！！！\n(￢_￢)',
-  '旅行者副本零掉落，旅行者深渊打不过，旅行者抽卡全保底，旅行者小保底必歪，旅行者找不到女朋友....',
-  '诅咒你在公共厕所大号没有纸!',
-  '你干嘛！',
-  '休息一下好不好',
-  '我生气了！咋瓦乐多!木大！木大木大！',
-  '你是不是喜欢我？',
-  '有完没完！',
-  '不准戳了！！！',
-  '朗达哟？',
-  '变态！',
-  '要被戳坏掉了>_<',
-  '大叔，你没工作的吗？一天天就知道戳我',
-  '哇啊啊！！真是的不可以戳戳\n>_<',
-  '别…别戳了……再戳…要戳坏掉了\n>_<',
-  '呐…呜~>_<',
-  '啊！请不要…碰我',
-  '嗯！轻一点！',
-  '唔呃！~',
-  '不要！！',
-  '不要……欺负…我',
-  '你干嘛……呃~痛！',
-  '哒咩！',
-  '可恶！',
-  '囊哒哟~',
-  '达咩！',
-  '呜哇！',
-  '你个坏蛋~',
-  '（摇头）',
-  '（后空翻）',
-  '（劈叉）',
-  '（惊醒）',
-  '（楞）',
-  '（眨眼）',
-  '？',
-  '？？',
-  '？？？',
-  '气气！',
-  '过分分！',
-  '走开啦！',
-  '੭ ᐕ)੭*⁾⁾',
-  '｀⌒´メ',
-  'o(´^｀)o',
-  '(。’▽’。)♡',
-  '(〟-_・)ﾝ?',
-  'Σ(°Д°;',
-  '⋟﹏⋞',
-  '◦˙▽˙◦',
-  'ξ( ✿＞◡❛)',
-  '_(:3 ⌒ﾞ)_',
-  '（╯‵□′）╯︵┴─┴',
-  'QAQ呜哇啊啊啊啊啊！',
-  '【旅行者命之座-1】',
-  '【旅行者保底次数+1】',
-  '【纠缠之缘-10】',
-  '【空月祝福-30】',
-  '【大冒险家的经验-100】',
-  '【摩拉-300w】',
-  '【原石-1600】',
-  '再戳，我就咬你！rua',
-  '真是的，不要再戳啦！！',
-  '唔，我生气了~~~',
-  '这个仇，我记下了！',
-  'QAQ请不要再戳啦！',
-  '哎，疼疼疼!',
-  '你还戳……']