+++ components/Path.js
@@ -50,7 +50,7 @@ export function getPath(_path) {
     ...path,
     /**
      * 获取路径
-     * @param {'_path'|'plugins'|'UC'|'apps'|'groupAdmin'|'botConfig'|'components'|'defSet'|'system'|'model'|'config'|'groupCfg'|'data'|'recycleBin'|'temp'|'QA'|'resources'|'unNecRes'|'img'|'example'|'sqtj'} rootDir 上级文件夹
+     * @param {'_path'|'plugins'|'UC'|'apps'|'groupAdmin'|'botConfig'|'components'|'defSet'|'system'|'model'|'config'|'groupCfg'|'data'|'recycleBin'|'temp'|'QA'|'resources'|'resdata'|'unNecRes'|'img'|'example'|'sqtj'} rootDir 上级文件夹
      * @param {string} basename 文件夹或文件basename
      * @returns 目标路径
      */